PPTP-291 Create Honesty Declaration page
This is a page that will allow us to integrate with the Generic
registration service.
In order to connect to the Generic Registration service we require a
/POST request to be made between our service and the generic
registration service.
This means we need to control that interaction via a /POST request to
our own service, as we should avoid /GET requests.