PPTP-292 Incorporated Entity Identification FE integration
This is just the integration between PPT FE and the IncorpID FE
microservice.

I have also organised the base spec folder, separating the `unit` and
`integration` base setup.

```
Outstanding work is to save the `journeyId` from the IncorpId FE
microservice in order to manage the statuses of each task.
Subsequent tickets will address the above.
```