PPTP-389 Change the feedback link construction
I mistakingly inserted the contact-frontend service on our lists of
microservice dependencies, when we do not need that, as we simply
require a HMRC base URI, as we did with the exit survey.

In order to provide the `backUrl` logic on the `contact-frontend`
service we will need to use the `platform.frontend.host` property which
is available on all environments, but we need to have a fallback
solution for local.