BAU - Attempt to load missing CSS on PPT Returns FE
Because we are sort of sharing the same URI between PPT registrations FE
and PPT returns FE, it could be that there is a confusion on what CSS to
load and probably PPT registrations FE is overriding the CSS config.

It works well locally because the domain (IP:PORT) is different and
there is no "confusion" there.

In QA the domain and start of URI is the same.

Ensuring that the bespoke PPT css partials are the last thing to be
loaded, to avoid them getting overwritten.