PPTP-893 Use 'InternalId' to save registration progress
At the beginning of Alpha phase we have used UTR to save the
registration user journey progress.
As we progressed through Alpha, we were told that two legal entities we
had to support woud not have an UTR.

This work is about replacing the enrolment key UTR with the User
Internal Id, which is always present in the Auth record and is unique to
this user.

This will allow us to continue to support the "Save And Come Back Later"
functionality.

Prior to this ticket we also had a UTR allowlist which will have to be
replaced with an email allowlist, which is a standard practice in other
HMRC services during Private Beta phase.

Subsequent commits, still part of PPTP-893  will refactor those classes
and the application config.