PPTP-993 : Integrate with GRS General Partnership Journey
Initially this only supports "general" partnerships. I've created the GrsIntegration trait in order to encapsulate the behaviour of GRS integration connectors (and updated the existing connectors to extends this). It would have been nice to have implemented the createJourney() and getDetails() in the base trait but my Scala Play Json was not up to the task; perhaps we should address this as a piece of technical debt in the future?

I have included updates to the TamperMonkey scripts for Partnership Identification Service feature switch configuration and to support this newly integrated journey.