PPTP-1017 : Capture PPT weight page refactor
Refactored to be consistent with the latest version in the prototype. In order to include all the necessary help text and remain compliant with the GDS guidelines, I had to remove use of the main page question as a text field heading and add the main question and help text as separate page pieces.