PPTP-838 Copy Returns Microservice CSS
On the MDTP platform if two microservices share the same URI, then the
microservice will pick up the static assets from  whatever is first on
the platform configuration.

One way around this is to copy the CSS from returns and the account
section to the registration microservice.

On our local environments we do not have this issue because the domain
includes the port, i.e, http://localhost:8503 and http://localhost:8505
and in this case the assets are pulled accordigly.