PPTP-1420 : Post Registration - Add New Group Member
We reuse logic from the registration flow by refactoring existing controllers into a base controller and having both the registration and the post registration controllers extending these. It was necessary to introduce a RegistrationUpdater with two concrete implementations: one for registration and the other for post registration amendment.

It's possible that further work is required in the AddGroupMemberController depending on the boundaries of the work done on PPTP-1538.