Pptp 1497 address capture refactor (#381)
We have created a new address capture component which we now use everywhere where we capture an address. This will ask the user whether the address is UK or non-UK and use either ALF or our own non-UK address compliant address capture page respectively. Configuration for the address capture and details of the captured address are stored in the local Mongo cache.