PPTP-1756 : Rework Registration Submitted Confirmation Page
Access to feature flag state migrated from JourneyRequest to AuthenticatedRequest. This was necessary since we do not want to use JourneyAction in the confirmation controller as we do not want a shell registration creating. This had a knock-on effect in a lot of tests which were creating AuthenticatedRequests and JourneyRequests.